テーブルを表示: Rename components/DataViewExpantionPanel.vue to components/index/_shared/DataView/ExpantionPanel.vue
